Add removefromarray functionality
todoItems and projects should have the ability to remove themselves from
the array they pushed themselves to and that ability should be added
to them as soon as they were pushed to an array. This will make it
easier later to remove themselves from arrays when delete functionality
is introduced.

A few DOM creation code will be added but they havent been polished.